+++ CHANGELOG.md
@@ -1,1 +1,4 @@
-- Fixed entire reagent bank being displayed sometimes
+- Added support for reagent bag
+- Added support for profession items
+- Added support for keyring (classic)
+- Added support for Dragonflight faction reputation